Add support for executing multiple Redis commands in a transaction
RedisConnection objects can now create Transactions that use that
particular connection to execute multiple queued commands in the context
of a Redis transaction.

RedisCommand objects can now .query() to execute immediately or .queue()
to execute later in the context of a transaction.

There are not currently many safeguards in place to ensure correct usage
of transactions, commands, connections, result objects, etc..